dedupe band/whiskers
